only consider predictions where departure_time is not null, and work on integrating with main page
