+++ lib/prediction_analyzer_web/templates/partials/header.html.eex
@@ -1,0 +1,7 @@
+<%= for mode_id <- [:subway, :commuter_rail] do %>
+  <%= content_tag :a, mode_string(mode_id), class: button_class(@conn, mode_id), href: Routes.accuracy_path(@conn, mode_id) %>
+<% end %>
+<div class="">
+  Show Dev Green results
+  <input type="checkbox" id="show-dev-green-check" checked>
+</div>
