+++ lib/prediction_analyzer_web/views/partials_view.ex
@@ -1,0 +1,3 @@
+defmodule PredictionAnalyzerWeb.PartialsView do
+  use PredictionAnalyzerWeb, :view
+end
